+++ src/Bootstrapper/Cake.Bootstrapper/InstallCmdlet.cs
@@ -3,7 +3,7 @@ using Autofac;
 using Cake.Bootstrapper.Installer;
 using Cake.Bootstrapper.Installer.GitIgnore;
 using Cake.Bootstrapper.Installer.NuGet;
-using Cake.Bootstrapper.Installer.Scripts;
+using Cake.Bootstrapper.Installer.Resources;
 
 namespace Cake.Bootstrapper
 {
@@ -12,6 +12,9 @@ namespace Cake.Bootstrapper
     {
         [Parameter]
         public string Source { get; set; }
+
+        [Parameter]
+        public SwitchParameter AppVeyor { get; set; }
 
         public override void RegisterDependencies(ContainerBuilder builder)
         {
@@ -23,17 +26,23 @@ namespace Cake.Bootstrapper
                 .As<INugetPackageVersionProber>()
                 .SingleInstance();
 
-            builder.RegisterType<ScriptCopier>().As<IScriptCopier>().SingleInstance();
+            builder.RegisterType<FileCopier>().As<IFileCopier>().SingleInstance();
 
             builder.RegisterType<GitIgnorePatcher>().As<IGitIgnorePatcher>().SingleInstance();
         }
 
         public override void SetCommandParameters(InstallCommand command)
         {
+            // Source
             if (!string.IsNullOrWhiteSpace(Source))
             {
-                // Set the source.
                 command.Source = Source;
+            }
+
+            // AppVeyor
+            if (AppVeyor.IsPresent)
+            {
+                command.AppVeyor = true;
             }
         }
     }