+++ src/Bootstrapper/Cake.Bootstrapper/Installer/GitIgnore/IGitIgnorePatcher.cs
@@ -4,6 +4,6 @@ namespace Cake.Bootstrapper.Installer.GitIgnore
 {
     public interface IGitIgnorePatcher
     {
-        void Patch(FilePath path);
+        bool Patch(FilePath path);
     }
 }