+++ src/Bootstrapper/Cake.Bootstrapper/Installer/InstallCommand.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Cake.Bootstrapper.Installer.GitIgnore;
 using Cake.Bootstrapper.Installer.NuGet;
-using Cake.Bootstrapper.Installer.Scripts;
+using Cake.Bootstrapper.Installer.Resources;
 using Cake.Bootstrapper.Net;
 using Cake.Bootstrapper.Runtime;
 using Cake.Core;
@@ -17,22 +17,23 @@ namespace Cake.Bootstrapper.Installer
         private readonly ICakeEnvironment _environment;
         private readonly ICakeLog _log;
         private readonly INuGetPackageConfigurationCreator _packageConfigCreator;
-        private readonly IScriptCopier _scriptCopier;
+        private readonly IFileCopier _fileCopier;
         private readonly IHttpDownloader _downloader;
         private readonly IGitIgnorePatcher _gitIgnorePatcher;
 
         public string Source { get; set; }
+        public bool AppVeyor { get; set; }
 
         public InstallCommand(IRuntime runtime, IFileSystem fileSystem, ICakeEnvironment environment,
             ICakeLog log, INuGetPackageConfigurationCreator packageConfigCreator,
-            IScriptCopier scriptCopier, IHttpDownloader downloader, IGitIgnorePatcher gitIgnorePatcher)
+            IFileCopier fileCopier, IHttpDownloader downloader, IGitIgnorePatcher gitIgnorePatcher)
         {
             _runtime = runtime;
             _fileSystem = fileSystem;
             _environment = environment;
             _log = log;
             _packageConfigCreator = packageConfigCreator;
-            _scriptCopier = scriptCopier;
+            _fileCopier = fileCopier;
             _downloader = downloader;
             _gitIgnorePatcher = gitIgnorePatcher;
 
@@ -51,7 +52,7 @@ namespace Cake.Bootstrapper.Installer
                 toolsDirectory.Create();
             }
 
-            // Download NuGet.exe              
+            // Download NuGet.exe.
             ReportProgress("Downloading NuGet executable...", 25);
             var nugetFilePath = toolsPath.MakeAbsolute(_environment).CombineWithFilePath("nuget.exe");
             if (!_fileSystem.Exist(nugetFilePath))
@@ -60,7 +61,7 @@ namespace Cake.Bootstrapper.Installer
                 _log.Information(" -> Downloaded NuGet executable.");
             }
 
-            // Create packages.config
+            // Create packages.config.
             ReportProgress("Generating NuGet package configuration...", 50);
             var packagesPath = toolsPath.MakeAbsolute(_environment).CombineWithFilePath("packages.config");
             if (!_fileSystem.Exist(packagesPath))
@@ -69,22 +70,34 @@ namespace Cake.Bootstrapper.Installer
                 _log.Information(" -> Generated NuGet package configuration.");
             }
 
-            // Copy bootstrapper script
+            // Copy bootstrapper script.
             ReportProgress("Copying bootstrapper script...", 75);
             var bootstrapperPath = new FilePath("build.ps1").MakeAbsolute(_environment);
             if (!_fileSystem.Exist(bootstrapperPath))
             {
-                _scriptCopier.Copy("build.ps1");
+                _fileCopier.Copy("build.ps1");
                 _log.Information(" -> Copied bootstrapper script.");
             }
 
-            // Copy build script
+            // Copy build script.
             ReportProgress("Preparing build script...", 85);
             var buildScriptPath = new FilePath("build.cake").MakeAbsolute(_environment);
             if (!_fileSystem.Exist(buildScriptPath))
             {
-                _scriptCopier.Copy("build.cake");
+                _fileCopier.Copy("build.cake");
                 _log.Information(" -> Copied build script.");
+            }
+
+            // Copy appveyor file.
+            if (AppVeyor)
+            {
+                ReportProgress("Copying AppVeyor configuration file...", 95);
+                var appVeyorPath = new FilePath("appveyor.yml").MakeAbsolute(_environment);
+                if (!_fileSystem.Exist(appVeyorPath))
+                {
+                    _fileCopier.Copy("appveyor.yml");
+                    _log.Information(" -> Copied AppVeyor configuration file.");
+                }
             }
 
             // Patch .gitignore file.
@@ -92,8 +105,10 @@ namespace Cake.Bootstrapper.Installer
             var gitIgnorePath = new FilePath(".gitignore").MakeAbsolute(_environment);
             if (_fileSystem.Exist(gitIgnorePath))
             {
-                _gitIgnorePatcher.Patch(gitIgnorePath);
-                _log.Information(" -> Patched .gitignore.");
+                if (_gitIgnorePatcher.Patch(gitIgnorePath))
+                {
+                    _log.Information(" -> Patched .gitignore.");   
+                }                
             }            
         }