+++ src/Bootstrapper/Cake.Bootstrapper/Installer/Resources/IFileCopier.cs
@@ -1,0 +1,7 @@
+namespace Cake.Bootstrapper.Installer.Resources
+{
+    public interface IFileCopier
+    {
+        bool Copy(string scriptName);
+    }
+}