+++ src/Bootstrapper/Cake.Bootstrapper/Installer/Scripts/IScriptCopier.cs
@@ -1,7 +1,0 @@
-namespace Cake.Bootstrapper.Installer.Scripts
-{
-    public interface IScriptCopier
-    {
-        bool Copy(string scriptName);
-    }
-}